requirements.txt
streamlit
requests
Pillow
python-dotenv
wolframalpha
plotly
groq
pandas
matplotlib